container/requirements.txt
apache-airflow[postgres]
pandas
numpy
scikit-learn
pyyaml
joblib